+++ lib/src/syn/v1/operator.rs
@@ -9,7 +9,7 @@ use nom::combinator::map;
 use nom::{
 	branch::alt,
 	bytes::complete::{tag, tag_no_case},
-	character::complete::{char, u16, u8},
+	character::complete::{char, u32, u8},
 	combinator::{cut, opt, value},
 };
 
@@ -151,7 +151,7 @@ pub fn knn_distance(i: &str) -> IResult<&str, Distance> {
 
 pub fn knn(i: &str) -> IResult<&str, Operator> {
 	let (i, _) = char('<')(i)?;
-	let (i, k) = u16(i)?;
+	let (i, k) = u32(i)?;
 	let (i, dist) = opt(knn_distance)(i)?;
 	let (i, _) = char('>')(i)?;
 	Ok((i, Operator::Knn(k, dist)))
